requirements.txt
nextcord[voice]
yt-dlp
python-dotenv
requests
json
re